deployment_package/requirements.txt
telethon==1.32.1
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
aiohttp==3.9.1
psutil==5.9.6
asyncio-mqtt==0.16.1
openai==1.3.7
